Log Timer and Api
Created an Api using Springboot with Java as the language and gradle as a library producer with groovy as the Domain Specific Language.

The Api perfroms task of creating TODO task and storing it as well as fetching it for later purpose.

Created an annotation of my own to calculate the time of execution of a method of the Api using custom build annotation.